use doubles

+++ simplexnoise.cpp
@@ -86,11 +86,11 @@ static const int simplex[64][4] = {
 };
 
 
-int fastfloor( const float x ) { return x > 0 ? (int) x : (int) x - 1; }
+int fastfloor( const double x ) { return x > 0 ? (int) x : (int) x - 1; }
 
-float dot( const int* g, const float x, const float y ) { return g[0]*x + g[1]*y; }
-float dot( const int* g, const float x, const float y, const float z ) { return g[0]*x + g[1]*y + g[2]*z; }
-float dot( const int* g, const float x, const float y, const float z, const float w ) { return g[0]*x + g[1]*y + g[2]*z + g[3]*w; }
+double dot( const int* g, const double x, const double y ) { return g[0]*x + g[1]*y; }
+double dot( const int* g, const double x, const double y, const double z ) { return g[0]*x + g[1]*y + g[2]*z; }
+double dot( const int* g, const double x, const double y, const double z, const double w ) { return g[0]*x + g[1]*y + g[2]*z + g[3]*w; }
 
 
 // Initialize permutation table with new pseudorandom values
@@ -106,14 +106,14 @@ void init_noise(int seed) {
 //
 // For each octave, a higher frequency/lower amplitude function will be added to the original.
 // The higher the persistence [0-1], the more of each succeeding octave will be added.
-float octave_noise_2d( const int octaves, const float persistence, const float scale, const float x, const float y ) {
-    float total = 0;
-    float frequency = scale;
-    float amplitude = 1;
+double octave_noise_2d( const int octaves, const double persistence, const double scale, const double x, const double y ) {
+    double total = 0;
+    double frequency = scale;
+    double amplitude = 1;
 
     // We have to keep track of the largest possible amplitude,
     // because each octave adds more, and we need a value in [-1, 1].
-    float maxAmplitude = 0;
+    double maxAmplitude = 0;
 
     for( int i=0; i < octaves; i++ ) {
         total += raw_noise_2d( x * frequency, y * frequency ) * amplitude;
@@ -131,14 +131,14 @@ float octave_noise_2d( const int octaves, const float persistence, const float s
 //
 // For each octave, a higher frequency/lower amplitude function will be added to the original.
 // The higher the persistence [0-1], the more of each succeeding octave will be added.
-float octave_noise_3d( const int octaves, const float persistence, const float scale, const float x, const float y, const float z ) {
-    float total = 0;
-    float frequency = scale;
-    float amplitude = 1;
+double octave_noise_3d( const int octaves, const double persistence, const double scale, const double x, const double y, const double z ) {
+    double total = 0;
+    double frequency = scale;
+    double amplitude = 1;
 
     // We have to keep track of the largest possible amplitude,
     // because each octave adds more, and we need a value in [-1, 1].
-    float maxAmplitude = 0;
+    double maxAmplitude = 0;
 
     for( int i=0; i < octaves; i++ ) {
         total += raw_noise_3d( x * frequency, y * frequency, z * frequency ) * amplitude;
@@ -156,14 +156,14 @@ float octave_noise_3d( const int octaves, const float persistence, const float s
 //
 // For each octave, a higher frequency/lower amplitude function will be added to the original.
 // The higher the persistence [0-1], the more of each succeeding octave will be added.
-float octave_noise_4d( const int octaves, const float persistence, const float scale, const float x, const float y, const float z, const float w ) {
-    float total = 0;
-    float frequency = scale;
-    float amplitude = 1;
+double octave_noise_4d( const int octaves, const double persistence, const double scale, const double x, const double y, const double z, const double w ) {
+    double total = 0;
+    double frequency = scale;
+    double amplitude = 1;
 
     // We have to keep track of the largest possible amplitude,
     // because each octave adds more, and we need a value in [-1, 1].
-    float maxAmplitude = 0;
+    double maxAmplitude = 0;
 
     for( int i=0; i < octaves; i++ ) {
         total += raw_noise_4d( x * frequency, y * frequency, z * frequency, w * frequency ) * amplitude;
@@ -181,7 +181,7 @@ float octave_noise_4d( const int octaves, const float persistence, const float s
 // 2D Scaled Multi-octave Simplex noise.
 //
 // Returned value will be between loBound and hiBound.
-float scaled_octave_noise_2d( const int octaves, const float persistence, const float scale, const float loBound, const float hiBound, const float x, const float y ) {
+double scaled_octave_noise_2d( const int octaves, const double persistence, const double scale, const double loBound, const double hiBound, const double x, const double y ) {
     return octave_noise_2d(octaves, persistence, scale, x, y) * (hiBound - loBound) / 2 + (hiBound + loBound) / 2;
 }
 
@@ -189,14 +189,14 @@ float scaled_octave_noise_2d( const int octaves, const float persistence, const 
 // 3D Scaled Multi-octave Simplex noise.
 //
 // Returned value will be between loBound and hiBound.
-float scaled_octave_noise_3d( const int octaves, const float persistence, const float scale, const float loBound, const float hiBound, const float x, const float y, const float z ) {
+double scaled_octave_noise_3d( const int octaves, const double persistence, const double scale, const double loBound, const double hiBound, const double x, const double y, const double z ) {
     return octave_noise_3d(octaves, persistence, scale, x, y, z) * (hiBound - loBound) / 2 + (hiBound + loBound) / 2;
 }
 
 // 4D Scaled Multi-octave Simplex noise.
 //
 // Returned value will be between loBound and hiBound.
-float scaled_octave_noise_4d( const int octaves, const float persistence, const float scale, const float loBound, const float hiBound, const float x, const float y, const float z, const float w ) {
+double scaled_octave_noise_4d( const int octaves, const double persistence, const double scale, const double loBound, const double hiBound, const double x, const double y, const double z, const double w ) {
     return octave_noise_4d(octaves, persistence, scale, x, y, z, w) * (hiBound - loBound) / 2 + (hiBound + loBound) / 2;
 }
 
@@ -205,7 +205,7 @@ float scaled_octave_noise_4d( const int octaves, const float persistence, const 
 // 2D Scaled Simplex raw noise.
 //
 // Returned value will be between loBound and hiBound.
-float scaled_raw_noise_2d( const float loBound, const float hiBound, const float x, const float y ) {
+double scaled_raw_noise_2d( const double loBound, const double hiBound, const double x, const double y ) {
     return raw_noise_2d(x, y) * (hiBound - loBound) / 2 + (hiBound + loBound) / 2;
 }
 
@@ -213,39 +213,39 @@ float scaled_raw_noise_2d( const float loBound, const float hiBound, const float
 // 3D Scaled Simplex raw noise.
 //
 // Returned value will be between loBound and hiBound.
-float scaled_raw_noise_3d( const float loBound, const float hiBound, const float x, const float y, const float z ) {
+double scaled_raw_noise_3d( const double loBound, const double hiBound, const double x, const double y, const double z ) {
     return raw_noise_3d(x, y, z) * (hiBound - loBound) / 2 + (hiBound + loBound) / 2;
 }
 
 // 4D Scaled Simplex raw noise.
 //
 // Returned value will be between loBound and hiBound.
-float scaled_raw_noise_4d( const float loBound, const float hiBound, const float x, const float y, const float z, const float w ) {
+double scaled_raw_noise_4d( const double loBound, const double hiBound, const double x, const double y, const double z, const double w ) {
     return raw_noise_4d(x, y, z, w) * (hiBound - loBound) / 2 + (hiBound + loBound) / 2;
 }
 
 
 
 // 2D raw Simplex noise
-float raw_noise_2d( const float x, const float y ) {
+double raw_noise_2d( const double x, const double y ) {
     // Noise contributions from the three corners
-    float n0, n1, n2;
+    double n0, n1, n2;
 
     // Skew the input space to determine which simplex cell we're in
-    float F2 = 0.5 * (sqrtf(3.0) - 1.0);
+    double F2 = 0.5 * (sqrtf(3.0) - 1.0);
     // Hairy factor for 2D
-    float s = (x + y) * F2;
+    double s = (x + y) * F2;
     int i = fastfloor( x + s );
     int j = fastfloor( y + s );
 
-    float G2 = (3.0 - sqrtf(3.0)) / 6.0;
-    float t = (i + j) * G2;
+    double G2 = (3.0 - sqrtf(3.0)) / 6.0;
+    double t = (i + j) * G2;
     // Unskew the cell origin back to (x,y) space
-    float X0 = i-t;
-    float Y0 = j-t;
+    double X0 = i-t;
+    double Y0 = j-t;
     // The x,y distances from the cell origin
-    float x0 = x-X0;
-    float y0 = y-Y0;
+    double x0 = x-X0;
+    double y0 = y-Y0;
 
     // For the 2D case, the simplex shape is an equilateral triangle.
     // Determine which simplex we are in.
@@ -256,10 +256,10 @@ float raw_noise_2d( const float x, const float y ) {
     // A step of (1,0) in (i,j) means a step of (1-c,-c) in (x,y), and
     // a step of (0,1) in (i,j) means a step of (-c,1-c) in (x,y), where
     // c = (3-sqrt(3))/6
-    float x1 = x0 - i1 + G2; // Offsets for middle corner in (x,y) unskewed coords
-    float y1 = y0 - j1 + G2;
-    float x2 = x0 - 1.0 + 2.0 * G2; // Offsets for last corner in (x,y) unskewed coords
-    float y2 = y0 - 1.0 + 2.0 * G2;
+    double x1 = x0 - i1 + G2; // Offsets for middle corner in (x,y) unskewed coords
+    double y1 = y0 - j1 + G2;
+    double x2 = x0 - 1.0 + 2.0 * G2; // Offsets for last corner in (x,y) unskewed coords
+    double y2 = y0 - 1.0 + 2.0 * G2;
 
     // Work out the hashed gradient indices of the three simplex corners
     int ii = i & 255;
@@ -269,21 +269,21 @@ float raw_noise_2d( const float x, const float y ) {
     int gi2 = perm[ii+1+perm[jj+1]] % 12;
 
     // Calculate the contribution from the three corners
-    float t0 = 0.5 - x0*x0-y0*y0;
+    double t0 = 0.5 - x0*x0-y0*y0;
     if(t0<0) n0 = 0.0;
     else {
         t0 *= t0;
         n0 = t0 * t0 * dot(grad3[gi0], x0, y0); // (x,y) of grad3 used for 2D gradient
     }
 
-    float t1 = 0.5 - x1*x1-y1*y1;
+    double t1 = 0.5 - x1*x1-y1*y1;
     if(t1<0) n1 = 0.0;
     else {
         t1 *= t1;
         n1 = t1 * t1 * dot(grad3[gi1], x1, y1);
     }
 
-    float t2 = 0.5 - x2*x2-y2*y2;
+    double t2 = 0.5 - x2*x2-y2*y2;
     if(t2<0) n2 = 0.0;
     else {
         t2 *= t2;
@@ -297,24 +297,24 @@ float raw_noise_2d( const float x, const float y ) {
 
 
 // 3D raw Simplex noise
-float raw_noise_3d( const float x, const float y, const float z ) {
-    float n0, n1, n2, n3; // Noise contributions from the four corners
+double raw_noise_3d( const double x, const double y, const double z ) {
+    double n0, n1, n2, n3; // Noise contributions from the four corners
 
     // Skew the input space to determine which simplex cell we're in
-    float F3 = 1.0/3.0;
-    float s = (x+y+z)*F3; // Very nice and simple skew factor for 3D
+    double F3 = 1.0/3.0;
+    double s = (x+y+z)*F3; // Very nice and simple skew factor for 3D
     int i = fastfloor(x+s);
     int j = fastfloor(y+s);
     int k = fastfloor(z+s);
 
-    float G3 = 1.0/6.0; // Very nice and simple unskew factor, too
-    float t = (i+j+k)*G3;
-    float X0 = i-t; // Unskew the cell origin back to (x,y,z) space
-    float Y0 = j-t;
-    float Z0 = k-t;
-    float x0 = x-X0; // The x,y,z distances from the cell origin
-    float y0 = y-Y0;
-    float z0 = z-Z0;
+    double G3 = 1.0/6.0; // Very nice and simple unskew factor, too
+    double t = (i+j+k)*G3;
+    double X0 = i-t; // Unskew the cell origin back to (x,y,z) space
+    double Y0 = j-t;
+    double Z0 = k-t;
+    double x0 = x-X0; // The x,y,z distances from the cell origin
+    double y0 = y-Y0;
+    double z0 = z-Z0;
 
     // For the 3D case, the simplex shape is a slightly irregular tetrahedron.
     // Determine which simplex we are in.
@@ -336,15 +336,15 @@ float raw_noise_3d( const float x, const float y, const float z ) {
     // a step of (0,1,0) in (i,j,k) means a step of (-c,1-c,-c) in (x,y,z), and
     // a step of (0,0,1) in (i,j,k) means a step of (-c,-c,1-c) in (x,y,z), where
     // c = 1/6.
-    float x1 = x0 - i1 + G3; // Offsets for second corner in (x,y,z) coords
-    float y1 = y0 - j1 + G3;
-    float z1 = z0 - k1 + G3;
-    float x2 = x0 - i2 + 2.0*G3; // Offsets for third corner in (x,y,z) coords
-    float y2 = y0 - j2 + 2.0*G3;
-    float z2 = z0 - k2 + 2.0*G3;
-    float x3 = x0 - 1.0 + 3.0*G3; // Offsets for last corner in (x,y,z) coords
-    float y3 = y0 - 1.0 + 3.0*G3;
-    float z3 = z0 - 1.0 + 3.0*G3;
+    double x1 = x0 - i1 + G3; // Offsets for second corner in (x,y,z) coords
+    double y1 = y0 - j1 + G3;
+    double z1 = z0 - k1 + G3;
+    double x2 = x0 - i2 + 2.0*G3; // Offsets for third corner in (x,y,z) coords
+    double y2 = y0 - j2 + 2.0*G3;
+    double z2 = z0 - k2 + 2.0*G3;
+    double x3 = x0 - 1.0 + 3.0*G3; // Offsets for last corner in (x,y,z) coords
+    double y3 = y0 - 1.0 + 3.0*G3;
+    double z3 = z0 - 1.0 + 3.0*G3;
 
     // Work out the hashed gradient indices of the four simplex corners
     int ii = i & 255;
@@ -356,28 +356,28 @@ float raw_noise_3d( const float x, const float y, const float z ) {
     int gi3 = perm[ii+1+perm[jj+1+perm[kk+1]]] % 12;
 
     // Calculate the contribution from the four corners
-    float t0 = 0.6 - x0*x0 - y0*y0 - z0*z0;
+    double t0 = 0.6 - x0*x0 - y0*y0 - z0*z0;
     if(t0<0) n0 = 0.0;
     else {
         t0 *= t0;
         n0 = t0 * t0 * dot(grad3[gi0], x0, y0, z0);
     }
 
-    float t1 = 0.6 - x1*x1 - y1*y1 - z1*z1;
+    double t1 = 0.6 - x1*x1 - y1*y1 - z1*z1;
     if(t1<0) n1 = 0.0;
     else {
         t1 *= t1;
         n1 = t1 * t1 * dot(grad3[gi1], x1, y1, z1);
     }
 
-    float t2 = 0.6 - x2*x2 - y2*y2 - z2*z2;
+    double t2 = 0.6 - x2*x2 - y2*y2 - z2*z2;
     if(t2<0) n2 = 0.0;
     else {
         t2 *= t2;
         n2 = t2 * t2 * dot(grad3[gi2], x2, y2, z2);
     }
 
-    float t3 = 0.6 - x3*x3 - y3*y3 - z3*z3;
+    double t3 = 0.6 - x3*x3 - y3*y3 - z3*z3;
     if(t3<0) n3 = 0.0;
     else {
         t3 *= t3;
@@ -391,28 +391,28 @@ float raw_noise_3d( const float x, const float y, const float z ) {
 
 
 // 4D raw Simplex noise
-float raw_noise_4d( const float x, const float y, const float z, const float w ) {
+double raw_noise_4d( const double x, const double y, const double z, const double w ) {
     // The skewing and unskewing factors are hairy again for the 4D case
-    float F4 = (sqrtf(5.0)-1.0)/4.0;
-    float G4 = (5.0-sqrtf(5.0))/20.0;
-    float n0, n1, n2, n3, n4; // Noise contributions from the five corners
+    double F4 = (sqrtf(5.0)-1.0)/4.0;
+    double G4 = (5.0-sqrtf(5.0))/20.0;
+    double n0, n1, n2, n3, n4; // Noise contributions from the five corners
 
     // Skew the (x,y,z,w) space to determine which cell of 24 simplices we're in
-    float s = (x + y + z + w) * F4; // Factor for 4D skewing
+    double s = (x + y + z + w) * F4; // Factor for 4D skewing
     int i = fastfloor(x + s);
     int j = fastfloor(y + s);
     int k = fastfloor(z + s);
     int l = fastfloor(w + s);
-    float t = (i + j + k + l) * G4; // Factor for 4D unskewing
-    float X0 = i - t; // Unskew the cell origin back to (x,y,z,w) space
-    float Y0 = j - t;
-    float Z0 = k - t;
-    float W0 = l - t;
+    double t = (i + j + k + l) * G4; // Factor for 4D unskewing
+    double X0 = i - t; // Unskew the cell origin back to (x,y,z,w) space
+    double Y0 = j - t;
+    double Z0 = k - t;
+    double W0 = l - t;
 
-    float x0 = x - X0; // The x,y,z,w distances from the cell origin
-    float y0 = y - Y0;
-    float z0 = z - Z0;
-    float w0 = w - W0;
+    double x0 = x - X0; // The x,y,z,w distances from the cell origin
+    double y0 = y - Y0;
+    double z0 = z - Z0;
+    double w0 = w - W0;
 
     // For the 4D case, the simplex is a 4D shape I won't even try to describe.
     // To find out which of the 24 possible simplices we're in, we need to
@@ -455,22 +455,22 @@ float raw_noise_4d( const float x, const float y, const float z, const float w )
     l3 = simplex[c][3]>=1 ? 1 : 0;
     // The fifth corner has all coordinate offsets = 1, so no need to look that up.
 
-    float x1 = x0 - i1 + G4; // Offsets for second corner in (x,y,z,w) coords
-    float y1 = y0 - j1 + G4;
-    float z1 = z0 - k1 + G4;
-    float w1 = w0 - l1 + G4;
-    float x2 = x0 - i2 + 2.0*G4; // Offsets for third corner in (x,y,z,w) coords
-    float y2 = y0 - j2 + 2.0*G4;
-    float z2 = z0 - k2 + 2.0*G4;
-    float w2 = w0 - l2 + 2.0*G4;
-    float x3 = x0 - i3 + 3.0*G4; // Offsets for fourth corner in (x,y,z,w) coords
-    float y3 = y0 - j3 + 3.0*G4;
-    float z3 = z0 - k3 + 3.0*G4;
-    float w3 = w0 - l3 + 3.0*G4;
-    float x4 = x0 - 1.0 + 4.0*G4; // Offsets for last corner in (x,y,z,w) coords
-    float y4 = y0 - 1.0 + 4.0*G4;
-    float z4 = z0 - 1.0 + 4.0*G4;
-    float w4 = w0 - 1.0 + 4.0*G4;
+    double x1 = x0 - i1 + G4; // Offsets for second corner in (x,y,z,w) coords
+    double y1 = y0 - j1 + G4;
+    double z1 = z0 - k1 + G4;
+    double w1 = w0 - l1 + G4;
+    double x2 = x0 - i2 + 2.0*G4; // Offsets for third corner in (x,y,z,w) coords
+    double y2 = y0 - j2 + 2.0*G4;
+    double z2 = z0 - k2 + 2.0*G4;
+    double w2 = w0 - l2 + 2.0*G4;
+    double x3 = x0 - i3 + 3.0*G4; // Offsets for fourth corner in (x,y,z,w) coords
+    double y3 = y0 - j3 + 3.0*G4;
+    double z3 = z0 - k3 + 3.0*G4;
+    double w3 = w0 - l3 + 3.0*G4;
+    double x4 = x0 - 1.0 + 4.0*G4; // Offsets for last corner in (x,y,z,w) coords
+    double y4 = y0 - 1.0 + 4.0*G4;
+    double z4 = z0 - 1.0 + 4.0*G4;
+    double w4 = w0 - 1.0 + 4.0*G4;
 
     // Work out the hashed gradient indices of the five simplex corners
     int ii = i & 255;
@@ -484,35 +484,35 @@ float raw_noise_4d( const float x, const float y, const float z, const float w )
     int gi4 = perm[ii+1+perm[jj+1+perm[kk+1+perm[ll+1]]]] % 32;
 
     // Calculate the contribution from the five corners
-    float t0 = 0.6 - x0*x0 - y0*y0 - z0*z0 - w0*w0;
+    double t0 = 0.6 - x0*x0 - y0*y0 - z0*z0 - w0*w0;
     if(t0<0) n0 = 0.0;
     else {
         t0 *= t0;
         n0 = t0 * t0 * dot(grad4[gi0], x0, y0, z0, w0);
     }
 
-    float t1 = 0.6 - x1*x1 - y1*y1 - z1*z1 - w1*w1;
+    double t1 = 0.6 - x1*x1 - y1*y1 - z1*z1 - w1*w1;
     if(t1<0) n1 = 0.0;
     else {
         t1 *= t1;
         n1 = t1 * t1 * dot(grad4[gi1], x1, y1, z1, w1);
     }
 
-    float t2 = 0.6 - x2*x2 - y2*y2 - z2*z2 - w2*w2;
+    double t2 = 0.6 - x2*x2 - y2*y2 - z2*z2 - w2*w2;
     if(t2<0) n2 = 0.0;
     else {
         t2 *= t2;
         n2 = t2 * t2 * dot(grad4[gi2], x2, y2, z2, w2);
     }
 
-    float t3 = 0.6 - x3*x3 - y3*y3 - z3*z3 - w3*w3;
+    double t3 = 0.6 - x3*x3 - y3*y3 - z3*z3 - w3*w3;
     if(t3<0) n3 = 0.0;
     else {
         t3 *= t3;
         n3 = t3 * t3 * dot(grad4[gi3], x3, y3, z3, w3);
     }
 
-    float t4 = 0.6 - x4*x4 - y4*y4 - z4*z4 - w4*w4;
+    double t4 = 0.6 - x4*x4 - y4*y4 - z4*z4 - w4*w4;
     if(t4<0) n4 = 0.0;
     else {
         t4 *= t4;
